+++ internal/creator/creator_test.go
@@ -1,0 +1,16 @@
+package creator
+
+import (
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestInit(t *testing.T) {
+	require.NotNil(t, tpl)
+	err := tpl.ExecuteTemplate(os.Stdout, "pack_readme.md", map[string]string{
+		"PackName": "foo",
+	})
+	require.NoError(t, err)
+}